Split typechecker scope management
