+++ docs/PHASE_PLAN.md
@@ -142,6 +142,8 @@ checked-in docs, tests, and commits only.
 - Expression checking now delegates match, conditional, and loop lowering
   through focused control-flow helpers, keeping pattern binding, exhaustiveness,
   and loop-control lowering behavior unchanged while shrinking the dispatcher.
+- Typechecker scope and import lookup methods now live in a dedicated
+  scope-management helper module instead of the broad typechecker root.
 - Resolver validation replay now collects expected declaration symbols,
   expected local symbols, import-validation state, and behavior-association
   replay tasks in one declaration pass before checking resolver-owned extras,
