+++ src/typechecker/scope_management.rs
@@ -1,0 +1,44 @@
+use super::*;
+
+impl TypeChecker {
+    pub(crate) fn push_scope(&mut self) {
+        self.scopes.push(Scope::new());
+    }
+
+    pub(crate) fn pop_scope(&mut self) {
+        self.scopes.pop();
+    }
+
+    pub(crate) fn define_var(&mut self, name: &str, ty: Type) {
+        self.define_var_with_mutability(name, ty, false);
+    }
+
+    pub(crate) fn define_var_with_mutability(&mut self, name: &str, ty: Type, mutable: bool) {
+        if let Some(scope) = self.scopes.last_mut() {
+            scope.vars.insert(name.to_string(), VarInfo { ty, mutable });
+        }
+    }
+
+    pub(crate) fn lookup_var(&self, name: &str) -> Option<Type> {
+        self.lookup_var_info(name).map(|info| info.ty.clone())
+    }
+
+    pub(crate) fn lookup_var_info(&self, name: &str) -> Option<&VarInfo> {
+        for scope in self.scopes.iter().rev() {
+            if let Some(info) = scope.vars.get(name) {
+                return Some(info);
+            }
+        }
+        None
+    }
+
+    pub(crate) fn is_import(&self, name: &str) -> bool {
+        self.imports.contains_key(name)
+    }
+
+    pub(crate) fn is_root_std_import(&self, name: &str) -> bool {
+        self.imports
+            .get(name)
+            .is_some_and(|path| path == &["std".to_string()] || path == &["@std".to_string()])
+    }
+}
